Issue #19: Support for Spark 2.3
